Added optional page table of contents
You can use the 'tableofcontents' front matter option to enable or
disable the display of a table of contents. Setting this option will
override the new site-wide parameter 'tableofcontentswordcount' which
will automatically add a table of contents on pages with more than
the specified number of words. Set to '0' to always show a table of
contents, or set to '-1' to never show a table of contents (the
default behaviour).